# src/common/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bookstore API",
    "description": "API for managing books",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3030"
    }
  ],
  "paths": {
...
    "/book": {
      "get": {
        "summary": "Get all books",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" }
        }
      },
      "post": {
        "summary": "Create a new book",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "author": { "type": "string" },
                  "year": { "type": "integer" }
                },
                "required": ["title", "author"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}